Installing Gulp into your Project
But now we can install Node packages into our project. The first, is gulp! So type, npm install gulp --save-dev. Wait, didn't we already install this? Well, the original command - with the -g gave us the global gulp executable. This time we're actually installing gulp into our project so other libraries can use it. Don't forget the --save-dev part. That says, 'download this into my project AND add an entry into package.json for it.'

Great! Flip back to your editor. The package.json has a new devDependencies section and we have a new node_modules directory with gulp in it:

 17 lines  package.json
{
  'name': 'gulp-knpu',
  'version': '0.0.0',
... lines 4 - 7
  'devDependencies': {
    'gulp': '^3.8.11'
  },
... lines 11 - 14
  'license': 'ISC'
}
In Composer terms, devDependencies is the require key in composer.json and node_modules is the vendor/ directory. Ok, we're rocking some Node!