Making the default Task Useful
Whenever we start working on a project, we'll want to run the sass task to initially process things and then watch for future stuff. Hmm, what if we made the default task do this for us? Add watch to the array.

Now just run gulp! It runs sass first and then starts watching for changes. That's really nice.