We only have one CSS file, and gosh, that's just not very realistic. So, let's create a second one. Call it, layout.scss. To keep the dinosaurs happy, let's give the body a background - a nice dark green background.
Ya know, because they're, sorta, dark green.

Go back and refresh! Ok, it doesn't work quite yet. But hey! We do have the watch task running in the background, and it's looking for any .scss file in that directory. So we should at least have a new layout.css file. Ah, there it is.