Yep, we're combining multiple files into one. That's pretty cool. But our main.css has an embarrassing amount of whitespace. Gross. Let's minify this.
Another plugin to the rescue! This one is called gulp-minify-css. And yea, there are a lot of plugins that can minify CSS. But this is a good one, and it's a superfriend with the sourcemaps plugin. They've even given us a nice install line, so I'll copy that, stop my watch task, and get it downloading:

npm install --save-dev gulp-minify-css