Next, go steal the require line and paste it on top.
Oh, and let's put a var before that.

And once again, we're going to use the trusty pipe() function to push things through minifyCSS()

This is looking really nice. Oh, and most of these functions - like minifyCSS() or sass() do take some options. So if you need to customize how things are minified, you can totally do that.

Ok, go back and run gulp!

gulp
And now main.css is a single line. BUT, through the power of sourcemaps, we still get the correct styles.scss line in the inspector.