A lot of times, I'll only minify my CSS when I'm deploying. Then locally, keep the whitespace for debugging. This matters a lot less now that we have sourcemaps, but I still want a way to control the minification.
Here's the goal: when I run gulp, I want it to not minify. But if I run gulp --production, I do want it to minify. Got it?

Installing gulp-util

To parse out that flag, the gulp-util(https://github.com/gulpjs/gulp-util) is the perfect tool. For some reason, it's missing from the Gulp plugins page, so just Google for it.

First, get it installed! Of course, that's:

npm install gulp-util --save-dev