Let's add a new config value called production and set that to git add .util.env.production.
Those two exclamations turn undefined into a proper false. It's a clever way to clean things up!

Conditional Statements in pipe()

Now all we need to do is add an if statement around the minifyCSS() line, right? Right? Um, actually it's not so easy.

The issue is with how the gulp stream works: you can't just stop in the middle of these pipes and keep going after an if statement. It just doesn't work like you'd think.

Instead, let's add some if logic right inside the pipe() that says if config.production, let's minifyCSS(), else, run this through a filter called util.noop.

This filter does absolutely nothing. Woo! Isn't that brilliant? So if we're not in production, the pipe() still happens, but no changes are made.

Moment of truth. First, just run gulp:

gulp
In this case, main.css is not minified. Go back, hit ctrl+c, and add the --production:

gulp --production
Goodbye whitespace! And in case you want to do anything else different in production, you've got a handy config value.