Instead, let's get to work. We need to pipe gulp through this plugin before any logic that might cause an error. So right after gulp.src, say pipe(plugins.plumber()).
Tip
The function(error) callback function was not shown in the video, but should be included. This prevents an issue when running gulp watch(https://knpuniversity.com/screencast/gulp/on-end-async-and-listeners#comment-2844811049).

And yea, that's it. Go back and get gulp back and running:

gulp
Let's mess up layout.scss again! This time, gulp does show us the error, it just doesn't die anymore. How nice! When we fix the error, it recompiles. Robots, get back to work!

 Tip
Plumber prevents gulp from throwing a proper error exit code. When building for production, you may want a proper error. If so, try using plumber() only in development.