Isolating the Styles Pipeline Now, how can we get Gulp to do all of this same logic, but dump out a new dinosaur.css file.
Start by creating a new variable called app. We'll use this as a place to store our own custom functions - including a nice new one called addStyle. Give this two arguments - the paths we want to process and the final filename to write. Next, copy the guts of the sass task into addStyle and make it dynamic: fill in paths on top, and filename instead of main.css