Renaming the Task to styles
And before we keep going, I think we can make another improvement. Our task is called sass. Let's change this to styles, because it's job is really to process styles, whether those are CSS or Sass files.

We also need to change that name on the default task. Ooops, and before restarting Gulp, also change the name on the watch task to styles.

Ok, now we can try things. ctrl+c to stop Gulp, and re-start it:

gulp
Yes, no errors! And the site still looks Tri-tastic! See what I did there?