scripts Task in Gulp
Create a new task called scripts.

Inside here, we're going to do almost the exact same stuff we do with our CSS. So let's copy the inside of addStyle and paste it in our task. Now we need to make a bunch of small adjustments.

First, get rid of a few things, like the sass filter and minifyCss. We will minify in a second, but we need to use a different tool.

Second, in src(), start with config.assetsDir then /js/main.js. We also need jQuery. That already lives inside the bower_components directory. So above main.js, add config.bowerDir - that new config sure is coming in handy - then /jquery/dist/jquery.js.

And to put the cherry on top, change the dest() to web/js and give concat() a filename - how about site.js.

Perfect. We're reusing all that good stuff we learned earlier!

Exit out of gulp with ctrl+c then just try running gulp scripts:

gulp scripts
Now in web/, we have a new site.js file and its map. So with almost no work, JS processing is alive!