Updating watch and default
Let's update the watch task so that the evil robots also rebuild things whenever we update a JS file. Copy the existing line. For the path, have it look for anything in the js directory recursively - so js/**/*.js. And when that happens, run scripts.

Add this to the default task too.

Let's exercise the whole system. Go back and just run gulp:

gulp
Great, it runs scripts and then the watch waits. To test that, go back and add a period to main.js, save and you can see that scripts ran again automatically.