Minify with gulp-uglify
We don't need the require line though, because gulp-plugins-require takes care of that for us. We can go straight to work. Copy the minifyCss line so that we have the cool --production flag behavior. Paste it and change things to say plugins.uglify().

Cool! Try it out first with just gulp:

gulp
This should give us the non-minified version. Excellent! Now ctrl+c that and try again with --production:

gulp --production
That file has been uglified! That kind of sounds like a Harry Potter spell. Uglify!