Dumping the rev-manifest.json File
We can't just update base.html.twig to use the new hashed name because it would re-break every time we changed the file. What we need is a map that says: 'Hey, main.css is actually called main-50d83f6c.css.' If we had that, we could use it inside our PHP code to rewrite the main.css in the base template to hashed version automatically. When the hashed name updates, the map would update, and so would our code.

And of course, the gulp-rev people thought of this! They call that map a 'manifest'. To get gulp-rev to create that for us, we need to ask it really nicely. At the end, add another pipe() to plugins.rev.manifest() and tell that where we want the manifest. Let's put it next to our assets at app/Resources/assets/rev-manifest.json.
As you'll see, this file doesn't need to be publicly accessible - our PHP code just needs to be able to read it.

There's one more interesting step: pipe() this into gulp.dest('.')

What?

What do Multiple dest()'s mean?

So far, we've always had one gulp.src() at the top and one gulp.dest() at the bottom, but you can have more. Our first gulp.dest() writes the CSS file. But once we pipe to plugins.rev.manifest(), the Gulp stream changes. Instead of being the CSS file, the manifest is now being passed through the pipes. So the last gulp.dest() just writes that file relative to the root directory.

Let me show you. Stop gulp and restart:

gulp
And there's our rev-manifest.json file:

{
  'main.css': 'main-50d83f6c.css'
}
It holds the map from main.css to its actual filename right now. It is missing dinosaur.css, but we'll fix that in a second.