Fixing the manifest base directory
But there's another problem I want to tackle first. In a second, we're going to put JavaScript paths into the manifest too. So I really need this to have the full public path - css/main.css - instead of just the filename.

So why does it just say main.css? Because when we call addStyle(), we pass in only main.css. This is passed to concat() and that becomes the path that's used by gulp-rev.

The fix is easy! Inside concat(), update it to css/ then the filename. That changes the filename that's inside the Gulp stream. To keep the file in the same spot, just take the css/ out of the gulp.dest() call

So nice: those two pipes work together to put the file in the same spot. Run gulp again:

gulp
Now, rev-manifest.json has the css/ prefix we need:

{
  'css/main.css': 'css/main-50d83f6c.css'
}