Merging Manifests
So why the heck doesn't my dinosaur.css show up here? The addStyle() function is called twice: once for main.css and once for dinosaur.css. But the second time, since the manifest file is already there, it does nothing. Unless, you pass an option called merge and set it to true.

Let's see if this fixed it! Re-run gulp:

gulp
Yes! The hard part is done - this is a perfect manifest file:

{
  'css/main.css': 'css/main-50d83f6c.css',
  'css/dinosaur.css': 'css/dinosaur-32046959.css'
}