Making the link href Dynamic
Phew! We're in the homestretch - the Gulp stuff is done. The only thing left is to make our PHP use the manifest file.

Since I'm in Twig, I'm going to invent a new filter called asset_version.

Let's make it do something! I already created an empty Twig extension file to get us started:

26 lines  src/AppBundle/Twig/AssetVersionExtension.php
<?php
namespace AppBundle\Twig;
class AssetVersionExtension extends \Twig_Extension{
  private {dollar_sign}appDir;
  public function __construct({dollar_sign}appDir){
    {dollar_sign}this->appDir = {dollar_sign}appDir;
  }
  public function getFilters(){
    return array(
    );
  }
  public function getName(){
    return 'asset_version';
  }
}
And I told Twig about this in my app/config/services.yml file:

 12 lines  app/config/services.yml
... lines 1 - 5
services:
  twig_asset_version_extension:
    class: AppBundle\Twig\AssetVersionExtension
    arguments: [%kernel.root_dir%]
    tags:
      - { name: twig.extension }
So, this Twig extension is ready to go! All we need to do is register this asset_version filter, which I'll do inside getFilters() with new \Twig_SimpleFilter('asset_version', ...) and we'll have it call a method in this class called getAssetVersion.

Below, I'll add that function. It'll be passed the  that we're trying to version. So for us, css/main.css.

Ok, our job is simple: open up rev-manifest.json, find the path, then return its versioned filename value. The path to that file is {dollar_sign}this->appDir - I've already setup that property to point to the app/ directory - then /Resources/assets/rev-manifest.json

With the power of TV, I'll magically add the next few lines. First, throw a clear exception if the file is missing. Next, open it up, decode the JSON, and set the map to an {dollar_sign}assets variable. Since the manifest file has the original filename as the key, let's throw one more exception if the file isn't in the map. I want to know when I mess up. And finally, return that mapped value!
So we give it css/main.css and it gives us the hashed filename.

Let's give it a shot! Take a deep breath and refresh. Victory! Our beautiful site is back - the hashed filename shows up in the source.