This is not a gulp plugin, so we need to go to the top and add var del = require('del') This library does what it sounds like - it helps delete files.
We need a way to clean up our generated files. That sounds like a perfect opportunity to add a new task! Call it clean.

Inside here, we want to remove everything that gulp generates. The first thing I can think of is the rev-manifest.json file. We don't need to clear this, but if you delete a CSS file, its last map value will still live here. So to keep only real files in this list, we might as well remove it entirely once in awhile.

To do that, use del.sync(). This means that our code will wait at this line until the file is actually deleted. The path to the manifest is further up. But hey, let's not duplicate! Instead, copy that path and reference a new config option called revManifestPath. Up top, I'll actually add that property to config.

Ok! Now just feed that to del.sync(). The other things we need to clean up are web/css/*, web/js/* and web/fonts/*

Great! Now, we could run this manually, but instead of doing that, add this to the beginning of the default task. So when I run gulp, it'll start by cleaning things up.

We have almost 10 things in web/css. Gulp, take out the trash.

gulp
My, what a tidy directory.