Ok, restart gulp!
gulp
Now check out js/. Very good - we've got site- the hash. And open up rev-manifest.json - it's got one new entry for this. The setup for CSS and JS files is no different - the maps are all getting put into the same file.

If we refresh now, we of course don't see our message. We're still including site.js. But since we already made our Twig plugin, we can pipe that path through asset_version. And that'll take care of everything.