Adding on('end') Listeners
Like with AJAX, each line returns something that acts like a Promise. That means, for any line, we can write on to add a listener for when this specific line actually finishes. When that happens, let's console.log('start '+filename).

Copy this and add another listener to the last line. Change the text to 'end'

Ok, run gulp!

gulp
Woh! When it said it finished 'styles', it really means it was done executing the styles task. But things really finish way later. In fact, they don't even start the process until later. And what's really crazy is that dinosaur.css starts before main.css, even though main is the first style we add.