Race Condition in the Manifest
There's a bug with our manifest file - a race condition! Ah gross! Because of the merge option, it opens up the manifest, reads the existing keys, updates one of them, then re-dumps the whole file.

For styles, the manifest file is opened twice for main.css and dinosaur.css. If one opens the file before the other finishes writing, when it writes, it'll run over the changes from the first.

How can we make the first addStyle finish before starting the second?

Using on to Control Order

It turns out the answer is easy. We can attach an end listener to any part of the Gulp stream. Return the stream from addStyle. Then in styles, attach an on('end') and only process dinosaur.css once the previous call is finished.

I know, it's ugly - we'll fix it, I promise! But let's see if it works:

gulp
Perfect! main.css starts and ends. Then dinosaur.css starts.