On top, add var Q = require('q')
To use it, create a pipeline variable and set it to new Pipeline(). Now, instead of calling app.addStyle() directly, call pipeline.add() with the same arguments. Now we can move dinosaur.css out of the nested callback and use pipeline.add() again. Woops, typo on pipeline!

pipeline.add is basically queuing those to be run. So at the end, call pipeline.run() and pass it the actual function it should call.

Behind the scenes, the Pipeline is doing what we did before: calling addStyle, waiting until it finishes, then calling addStyle again.

Try it!

gulp
Cool - we've got the same ordering.