Look at the default task. The array defines task dependencies: Gulp runs each of these first, waits for them to finish, and then executes the callback for the task, if there is one. And based on the output, it looks like it runs them in order: clean, then styles, then scripts.
Is that true?

There is no Order to Dependent Tasks

Log a message once fonts is done And also add a message right when the watch task starts

The default task defines fonts then watch, and I want to see if that order matters.

Ok, try it!

gulp
It exploded! It say we're calling on on something undefined. This happens with our code because up in app.copy, we're not returning the stream. So yea, that would be undefined