We don't need the fonts dependency, so take it off. And remove the logging.
So if we should always return a stream or promise, how can we do that for styles? It doesn't have a single stream - it has two that are combined into the pipeline. We need to wait until both of them are finished.
Oh, the answer is so nice: just return pipeline.run().

This isn't magic. I wrote the Pipeline code, and I made run() return a Promise that resolves once everything is done. And if you know anything about promises, the guts should make sense to you. But if you have questions, just ask in the comments.

Make sure we didn't break anything.

gulp
Yep, it all still looks great. So if you eventually need to create a task that's dependent on styles or scripts finishing first, it'll work.